Move AttackGlobals to core
