+++ core/AttackGlobals.py
