+++ tests/battle/test_battle_creation.py
@@ -1,6 +1,7 @@
 import pytest
 
-from ...AttackGlobals import MULTIPLIER, MULTIPLIER_DEFAULT
+from core.AttackGlobals import MULTIPLIER, MULTIPLIER_DEFAULT
+
 from ...Battle import Battle, BattleCog
 from ...Cog import Cog, get_random_cog
 from ...Toon import Toon
